Корректирует работу функций
